requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pymongo==4.6.0
motor==3.3.2
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
statsmodels==0.14.0
scikit-learn==1.3.2
prophet==1.1.5
langchain==0.1.1
langchain-openai==0.0.5
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6
